# Conflicts:
#	src/main/java/gov/cabinetofice/gapuserservice/service/user/OneLoginUserService.java
#	src/main/java/gov/cabinetofice/gapuserservice/util/HelperUtils.java
#	src/main/java/gov/cabinetofice/gapuserservice/web/LoginControllerV2.java
#	src/main/java/gov/cabinetofice/gapuserservice/web/UserController.java
#	src/test/java/gov/cabinetofice/gapuserservice/web/LoginControllerV2Test.java
#	src/test/java/gov/cabinetofice/gapuserservice/web/UserControllerTest.java

+++ src/main/java/gov/cabinetofice/gapuserservice/web/UserController.java
@@ -8,6 +8,7 @@ import gov.cabinetofice.gapuserservice.exceptions.InvalidRequestException;
 import gov.cabinetofice.gapuserservice.model.User;
 import gov.cabinetofice.gapuserservice.service.DepartmentService;
 import gov.cabinetofice.gapuserservice.service.RoleService;
+import gov.cabinetofice.gapuserservice.service.jwt.impl.CustomJwtServiceImpl;
 import gov.cabinetofice.gapuserservice.service.SecretAuthService;
 import gov.cabinetofice.gapuserservice.service.user.OneLoginUserService;
 import jakarta.servlet.http.Cookie;
@@ -21,6 +22,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 import static gov.cabinetofice.gapuserservice.util.HelperUtils.getCustomJwtCookieFromRequest;
 
@@ -33,10 +35,24 @@ public class UserController {
     private final OneLoginUserService oneLoginUserService;
     private final DepartmentService departmentService;
     private final RoleService roleService;
+    private final CustomJwtServiceImpl jwtService;
     private final SecretAuthService secretAuthService;
 
     @Value("${jwt.cookie-name}")
     public String userServiceCookieName;
+
+    @GetMapping("/userFromJwt")
+    public ResponseEntity<UserDto> getUserFromJwt(HttpServletRequest httpRequest) {
+        if (!roleService.isSuperAdmin(httpRequest)) {
+            throw new ForbiddenException();
+        }
+        Optional<User> user = jwtService.getUserFromJwt(httpRequest);
+        if(user.isEmpty()){
+            throw new InvalidRequestException("Could not get user from jwt");
+        }
+
+        return ResponseEntity.ok(new UserDto(user.get()));
+    }
 
     @GetMapping("/isSuperAdmin")
     public ResponseEntity<String> isSuperAdmin(HttpServletRequest httpRequest) {
@@ -100,6 +116,17 @@ public class UserController {
         if (!roleService.isSuperAdmin(httpRequest)) {
             throw new ForbiddenException();
         }
+
+        boolean isARequestToBlockUser = roleIds.size() == 0;
+        Optional<User> user = jwtService.getUserFromJwt(httpRequest);
+
+        if(user.isEmpty()){
+            throw new InvalidRequestException("Could not get user from jwt");
+        }
+        if (isARequestToBlockUser && id.equals(user.get().getGapUserId())){
+            throw new UnsupportedOperationException("You can't block yourself");
+        }
+
         oneLoginUserService.updateRoles(id, roleIds);
         return ResponseEntity.ok("success");
     }
@@ -110,8 +137,17 @@ public class UserController {
             throw new ForbiddenException();
         }
         final Cookie customJWTCookie = getCustomJwtCookieFromRequest(httpRequest, userServiceCookieName);
+        Optional<User> user = jwtService.getUserFromJwt(httpRequest);
+        if(user.isEmpty()){
+            throw new InvalidRequestException("Could not get user from jwt");
+        }
+        if(user.get().getGapUserId().equals(id)) {
+            throw new UnsupportedOperationException("You can't delete yourself");
+        }
 
         oneLoginUserService.deleteUser(id, customJWTCookie.getValue());
         return ResponseEntity.ok("success");
+
     }
 }
+
